Assignments/Week-7/S7/requirements.txt
pydantic>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.21.0
requests>=2.26.0
youtube-transcript-api>=0.6.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
fastapi>=0.68.0
uvicorn>=0.15.0 